requirements.txt
appdirs==1.4.4
apptools==5.2.0
asttokens==2.2.1
attrs==22.2.0
audioread==3.0.0
backcall==0.2.0
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.2
ConfigArgParse==1.5.3
configobj==5.0.8
contourpy==1.0.7
cycler==0.11.0
dash==2.8.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.6
decorator==5.1.1
dill==0.3.6
envisage==6.1.1
executing==1.2.0
fastjsonschema==2.16.3
fire==0.5.0
Flask==2.2.3
fonttools==4.38.0
idna==3.4
imageio==2.26.0
ipykernel==6.21.2
ipython==8.10.0
ipywidgets==8.0.4
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
lazy_loader==0.1
librosa==0.10.0.post2
llvmlite==0.39.1
MarkupSafe==2.1.2
matplotlib==3.7.0
matplotlib-inline==0.1.6
mayavi==4.8.1
msgpack==1.0.5
nbformat==5.5.0
nest-asyncio==1.5.6
networkx==3.0
numba==0.56.4
numpy==1.23.5
open3d==0.16.0
opencv-python==4.7.0.68
packaging==23.0
pandas==1.5.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.0.0
plotly==5.13.1
pooch==1.6.0
prompt-toolkit==3.0.37
protobuf==4.22.1
psutil==5.9.4
pure-eval==0.2.2
pycparser==2.21
pydub==0.25.1
pyface==7.4.4
Pygments==2.14.0
PyOpenGL==3.1.6
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
pyqtgraph==0.13.2
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2022.7.1
PyWavelets==1.4.1
pywin32==305
pyzmq==25.0.0
requests==2.28.2
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.13.0
shapely==2.0.1
six==1.16.0
soundfile==0.12.1
soxr==0.3.4
stack-data==0.6.2
tenacity==8.2.1
termcolor==2.2.0
threadpoolctl==3.1.0
tifffile==2023.2.28
tornado==6.2
tqdm==4.66.1
traitlets==5.9.0
traits==6.4.1
traitsui==7.4.3
typing_extensions==4.5.0
urllib3==1.26.15
vtk==9.2.6
wcwidth==0.2.6
Werkzeug==2.2.3
widgetsnbextension==4.0.5
wincertstore==0.2
